[Finish]7d

+++ db/migrate/20220430101416_add_star_to_books.rb
@@ -1,5 +1,5 @@
 class AddStarToBooks < ActiveRecord::Migration[6.1]
   def change
-    add_column :books, :star, :integer
+    add_column :books, :star, :float
   end
 end
